--- api/opentrons/server/helpers.py
@@ -1,8 +1,0 @@
-import sys
-
-
-def get_frozen_root():
-    """
-    :return: Returns app path when app is packaged by pyInstaller
-    """
-    return sys._MEIPASS if getattr(sys, 'frozen', False) else None
