--- api/opentrons/server/robot_container.py
@@ -1,1 +1,0 @@
-
