+++ api/opentrons/session/__init__.py
@@ -1,0 +1,8 @@
+from .session import Session, SessionManager
+from .notifications import Notifications
+
+__all__ = [
+    Session,
+    SessionManager,
+    Notifications
+]
